app/bot/requirements.txt
delta-rest-client>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-telegram-bot>=20.0
flask>=3.0.0
ta>=0.11.0
colorama>=0.4.6
pytz==2023.3